requirements.txt
streamlit
requests
python-pptx